Implemented rol, rolw, ror, and rorw RISC-V instruction emulation
