+++ src/Archs/RiscV/Zarem.RiscV.Emulator/Interpret/RiscVInstructionServiceTable.BitManipulation.cs
@@ -1,6 +1,7 @@
 ﻿// Avishai Dernis 2026
 
 using System.Numerics;
+using Zarem.Emulator.Models.Logic;
 using Zarem.RiscV.Emulator.Interpret;
 using Zarem.RiscV.Emulator.Machine.Enums;
 using Zarem.RiscV.Models.Instructions;
@@ -10,7 +11,7 @@ namespace Zarem.Emulator.Models;
 
 public partial class RiscVInstructionServiceTable<T, TFloat, TSigned>
 {
-    private static RiscVTrap BitCountSignExtend<T2, T2Signed>(RiscVInterpretCpu<T, TFloat> cpu, RiscVInstruction inst, bool compressed, out RiscVExecution<T> exec)
+    private static RiscVTrap BitCountSignExtendRotate<T2, T2Signed>(RiscVInterpretCpu<T, TFloat> cpu, RiscVInstruction inst, bool compressed, out RiscVExecution<T> exec)
         where T2 : unmanaged, IBinaryInteger<T2>, IUnsignedNumber<T2>
         where T2Signed : unmanaged, IBinaryInteger<T2Signed>, ISignedNumber<T2Signed>
     {
@@ -21,7 +22,7 @@ public partial class RiscVInstructionServiceTable<T, TFloat, TSigned>
             FunctRS2Code.PopulationCount => AluR<CpopLogic<T2>, T2>(cpu, inst, compressed, out exec),
             FunctRS2Code.SignExtendByte => AluR<Sext<T2, T2Signed, sbyte>, T2>(cpu, inst, compressed, out exec),
             FunctRS2Code.SignExtendHalfword => AluR<Sext<T2, T2Signed, short>, T2>(cpu, inst, compressed, out exec),
-            _ => IllegalInstruction(cpu, inst, compressed, out exec),
+            _ => AluR<RolLogic<T2>, T2>(cpu, inst, compressed, out exec),
         };
     }
 }
