+++ tests/Archs/RiscV/Test.RiscV/Emulator/RiscVEmulatorTestDataSourceAttribute.cs
@@ -232,6 +232,8 @@ public class RiscVEmulatorTestDataSourceAttribute : EmulatorTestDataSourceAttrib
                 yield return [new RiscVEmulatorTestCase<T>(config, "clz a0, s8", T.LeadingZeroCount(T.CreateTruncating(S8)))];
                 yield return [new RiscVEmulatorTestCase<T>(config, "ctz a0, s8", T.TrailingZeroCount(T.CreateTruncating(S8)))];
                 yield return [new RiscVEmulatorTestCase<T>(config, "cpop a0, s8", T.PopCount(T.CreateTruncating(S8)))];
+                yield return [new RiscVEmulatorTestCase<T>(config, "rol a0, t3, s3", T.RotateLeft(T.CreateTruncating(40), 4))];
+                yield return [new RiscVEmulatorTestCase<T>(config, "ror a0, t3, s3", T.RotateRight(T.CreateTruncating(40), 4))];
                 yield return [new RiscVEmulatorTestCase<T>(config, "sext.b a0, a5", T.CreateTruncating(TSigned.CreateTruncating((sbyte)0xF8)))];
 
                 yield return [new RiscVEmulatorTestCase<T>(config, "andn a0, s8, s9", T.CreateTruncating(S8) & ~T.CreateTruncating(S9))];
